extract osUpdate proposal to new file to fix other proposal pages from breaking
